JD_maker/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "groq>=1.2.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
